Proxy: escape ';' in URI paths
This fixes a security hole when used with many Java application servers,
which treat ';' as a special character in paths.  See [1] for more
details.

This changes how NGINX constructs requests for its backends.  Something
like:

    location /fake1/ {
        proxy_pass http://127.0.0.1:8000/fake2/;
    }

will now escape semicolons in the provided URL, provided that the
original URL had any %-escapes.

To prevent such bypasses, configurations can use:

    if ($request_uri !~ "^/[^;?#]*(?:\?|$)") {
        return 400 "Path parameters not allowed";
    }

This is ultimately a flaw in NGINX, not the backend servers: it is a
client error to make an HTTP request with a URI that contains a path
with a character that is not valid in paths, and as the client NGINX is
responsible for only making valid requests.

This is not a complete fix.  A complete fix would require blocking %2F
in paths (HAProxy and NGINX disagree on its meaning) and blocking
unescaped semicolons in paths (NGINX and Java app servers disagree on
their meaning).

This patch allows closing the remaining bypasses by doing the following
in configuration files:

1. Do not use regular expressions in paths unless careful.
2. Ensure all location blocks are either "location = /something" (note
   the = operator) or location /something/ (trailing slash).
3. Include the following at server level:

    if ($request_uri !~ "^(?<non_normalized_path>/[^;?#]*)(?:\?|$)") {
        return 400 "Path parameters not allowed";
    }

    if ($non_normalized_path ~ "%2[EeFf]|//|/\.\.?(?:\?|/|$)") {
        return 400 "Path has escaped . or /, repeated /, or . or .. component";
    }

This checks that path components are not empty and are not "." or "..".
It also checks that there are no path parameters and no escaped "." or
"/" characters.  Paths that meet these requirements will be interpreted
consistently by all web servers I am aware of.

One can consider this snippet to be licensed under the MIT-0 license (no
attribution required, use for any purpose allowed with no restrictions).

Without this patch to NGINX, normalized URLs submitted to backends would
have unescaped semicolons even if the URL submitted to NGINX escaped the
semicolon.  Preventing the bypass would therefore require blocking
*escaped* semicolons, which should not be required.

[1]: https://i.blackhat.com/us-18/Wed-August-8/us-18-Orange-Tsai-Breaking-Parser-Logic-Take-Your-Path-Normalization-Off-And-Pop-0days-Out-2.pdf